Extract useRelationEntities from SingleEntitySelectMenuItems
